CRUD Exercises

+++ Team_03/02_Source/01_Source_Code/BE/services/fit-service/src/main/java/com/hcmus/fitservice/model/Exercise.java
@@ -31,7 +31,7 @@ public class Exercise {
 
     @NotNull
     @Column(name = "calories_burned_per_minute", nullable = false)
-    private Double caloriesBurnedPerMinute;
+    private Integer caloriesBurnedPerMinute;
 
     @Column(name = "user_id")
     private UUID userId;
